# storage/api-docs/api-docs.yaml
...
openapi: 3.0.0
info:
  title: Development Challenge for PHP
  description: This is a sample API documentation for the Development Challenge for
    PHP
  version: 1.0.0
paths:
...
  "/api/tvshows":
    get:
      tags:
      - Tv Shows
      summary: Display a listing of the tv shows
      operationId: 71f677ed5bcb2d9d77bd04c0bc99db55
      responses:
        '200':
          description: A listing of the tv show
          content:
            application/json:
              schema:
                properties:
                  tvShows:
                    type: object
                type: object
    post:
      tags:
      - Tv Shows
      summary: Store a newly created tv show in storage
      operationId: 23de601ec5ea85096e5f0884a58ee543
      requestBody:
        description: Pass tv show data
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                release_date:
                  type: string
                rating:
                  type: integer
                genre:
                  type: string
                director_id:
                  type: integer
                age_classification_id:
                  type: integer
              type: object
      responses:
        '200':
          description: Tv show created successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  tvShow:
                    "$ref": "#/components/schemas/TvShow"
                type: object
        '422':
          description: Invalid data passed
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  errors:
                    type: object
                type: object
  "/api/tvshows/{tvShow}":
    get:
      tags:
      - Tv Shows
      summary: Display the specified tv show
      operationId: 3d178d66c304f972efaf3bb29c0b9c69
      responses:
        '200':
          description: Display the specified tv show
          content:
            application/json:
              schema:
                properties:
                  tvShow:
                    type: object
                type: object
        '404':
          description: Tv Show not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
    put:
      tags:
      - Tv Shows
      summary: Update the specified tv show in storage
      operationId: 4816b509da8eefb1f0c55a185f4f291d
      parameters:
      - name: tvShow
        in: path
        description: TV Show ID
        required: true
        schema:
          type: integer
      requestBody:
        description: Pass tv show data
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                release_date:
                  type: string
                rating:
                  type: integer
                genre:
                  type: string
                director_id:
                  type: integer
                age_classification_id:
                  type: integer
              type: object
      responses:
        '200':
          description: Tv show updated successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  tvShow:
                    "$ref": "#/components/schemas/TvShow"
                type: object
        '404':
          description: Tv Show not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
        '422':
          description: Invalid data passed
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  errors:
                    type: object
                type: object
    delete:
      tags:
      - Tv Shows
      summary: Remove the specified tv show from storage
      operationId: 8a06770cf7fd9dc22671c06e7f66ce48
      responses:
        '200':
          description: Tv Show deleted successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  tvShow:
                    type: object
                type: object
        '404':
          description: Tv Show not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
...
components:
  schemas:
    GenderEnum:
      description: 'Gender options: Male, Female'
      type: string
      enum:
      - M
      - F
    GenreEnum:
      description: 'Genre options: Action, Comedy, Drama, Fantasy, Horror, Mystery,
        Romance, Science Fiction, Thriller'
      type: string
      enum:
      - AC
      - CO
      - DR
      - FA
      - HO
      - MY
      - RO
      - SF
      - TH
...
    AgeClassification:
      title: AgeClassification
      description: AgeClassification model
      properties:
        id:
          description: AgeClassification id
          type: integer
          example: '1'
        code:
          description: AgeClassification code
          type: string
          example: G
        description:
          description: AgeClassification description
          type: string
          example: General Audiences
      type: object
    Director:
      title: Director
      description: Director model
      properties:
        id:
          description: Director id
          type: integer
          example: '1'
        name:
          description: Director name
          type: string
          example: John Doe
        birth_date:
          description: Director birth date
          type: string
          format: date
          example: '1990-01-01'
        gender:
          "$ref": "#/components/schemas/GenderEnum"
      type: object
...
    TvShow:
      title: TvShow
      description: TvShow model
      properties:
        id:
          description: TvShow id
          type: integer
          example: '1'
        title:
          description: TvShow title
          type: string
          example: The Witcher
        release_date:
          description: TvShow release date
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        synopsis:
          description: TvShow synopsis
          type: string
          example: The Witcher is an American fantasy drama streaming television series
            produced by Lauren Schmidt Hissrich. It is based on the book series of
            the same name by Polish writer Andrzej Sapkowski.
        rating:
          description: TvShow rating
          type: integer
          example: '5'
        genre:
          "$ref": "#/components/schemas/GenreEnum"
        director:
          "$ref": "#/components/schemas/Director"
        age_classification:
          "$ref": "#/components/schemas/AgeClassification"
        seasons:
          description: TvShow seasons
          type: array
          items:
            "$ref": "#/components/schemas/TvShowSeason"
      type: object
...
    TvShowSeason:
      title: TvShowSeason
      description: TvShow Season model
      required:
      - season_number
      - tv_show_id
      properties:
        id:
          description: TvShow Season id
          type: integer
          example: '1'
        season_number:
          description: TvShow Season number
          type: integer
          example: '1'
        tv_show_id:
          description: TvShow Season tv show id
          type: integer
          example: '1'
        created_at:
          description: TvShow Season created at
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updated_at:
          description: TvShow Season updated at
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        deleted_at:
          description: TvShow Season deleted at
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...